Modularize: Change dist/es to dist/esm
