+++ rollup.config.cjs
@@ -38,7 +38,7 @@ module.exports = defineConfig([
     input: 'src/index.ts',
     output: [
       {
-        dir: 'dist/es',
+        dir: 'dist/esm',
         format: 'es',
         preserveModules: true,
         preserveModulesRoot: 'src',
@@ -48,7 +48,7 @@ module.exports = defineConfig([
     ],
     external: [/node_modules/, /@hebcal/],
     plugins: [
-      typescript({...tsOptions, outDir: 'dist/es'}),
+      typescript({...tsOptions, outDir: 'dist/esm'}),
       commonjs(),
       nodeResolve(),
       json({compact: true, preferConst: true}),
